requirements.txt
atproto==0.0.64
html2text==2025.4.15
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
